変更: - `api/src/utils/timezone.ts` のインラインテストを削除し、`formatToTimezone`/`formatToISODate`/`isValidTimezone` をエクスポート化。 - `api/src/utils/timezone.test.ts` を追加し、元のインラインテスト相当のケースを外部テストとして移設。
テスト:
- `cd api && pnpm run format`（OK、変更なし）
- `cd api && pnpm exec tsc --noEmit`（失敗：`api/src/config/index.ts` の未使用 `@ts-expect-error`、`api/src/routes/seed.ts` の `json.data` 未定義警告、`api/src/services/markAsUnread.test.ts` の型不一致）
- `cd api && pnpm run test`（成功、意図したエラーログ/コンソール警告が一部出力）

補足:
- 無効タイムゾーンや不正日付のテストで `console.warn` が出ますが、挙動確認のための想定ログです。

次のアクション候補:
1. `api/src/config/index.ts` の未使用 `@ts-expect-error` を削除。
2. `api/src/routes/seed.ts` で `json.data` の未定義ガードを追加。
3. `api/src/services/markAsUnread.test.ts` の型定義を `IBookmarkRepository` に合わせて修正。
